[CCUBE-2148][ZZ] refactor: move forced mode hook
