+++ src/theme/utils/index.ts
@@ -3,4 +3,5 @@ export * from "./font";
 export * from "./token-resolver";
 export * from "./use-apply-styles";
 export * from "./use-design-token";
+export * from "./use-forced-mode-token";
 export * from "./use-media-query";
